cleanup, especially separated evaluator and stratified splitting
