--- src/evaluate_bias_experiments.py
@@ -1,83 +1,0 @@
-import argparse
-import copy
-from pathlib import Path
-
-import yaml
-
-from src.datasets.helper import DatasetName
-from src.trainers.bias_experiment_standard_split import BiasExperimentStandardSplit
-from src.utils.loader import Loader
-
-my_parser = argparse.ArgumentParser(description="Experiments for the PASSION paper.")
-my_parser.add_argument(
-    "--config_path",
-    type=str,
-    required=True,
-    help="Path to the config yaml.",
-)
-my_parser.add_argument(
-    "--exp1",
-    action="store_true",
-    help="If the experiment 1, i.e. differential diagnosis, should be run.",
-)
-my_parser.add_argument(
-    "--exp2",
-    action="store_true",
-    help="If the experiment 2, i.e. detecting impetigo, should be run.",
-)
-# todo: fix
-my_parser.add_argument(
-    "--exp3",
-    action="store_true",
-    help="If the experiment 3, i.e. generalization collection centers, should be run.",
-)
-my_parser.add_argument(
-    "--exp4",
-    action="store_true",
-    help="If the experiment 4, i.e. generalization age groups, should be run.",
-)
-my_parser.add_argument(
-    "--append_results",
-    action="store_true",
-    help="If the results should be appended to the existing df (needs to be used with care!)",
-)
-my_parser.add_argument(
-    "--eval_only",
-    action="store_true",
-    help="If set, will only run evaluation without training. Requires saved checkpoint.",
-)
-args = my_parser.parse_args()
-
-if __name__ == "__main__":
-    # load config yaml
-    args.config_path = Path(args.config_path)
-    if not args.config_path.exists():
-        raise ValueError(f"Unable to find config yaml file: {args.config_path}")
-    config = yaml.load(open(args.config_path, "r"), Loader=Loader)
-    # overall parameters used for all datasets
-    log_wandb = config.pop("log_wandb")
-
-    model = "imagenet"
-    if args.exp1:
-        trainer = BiasExperimentStandardSplit(
-            dataset_name=DatasetName.PASSION,
-            config=config,
-            SSL_model=model,
-            append_to_df=args.append_results,
-            log_wandb=log_wandb,
-            add_info="conditions",
-        )
-        trainer.evaluate()
-
-    if args.exp2:
-        _config = copy.deepcopy(config)
-        _config["dataset"]["passion"]["label_col"] = "IMPETIGO"
-        trainer = BiasExperimentStandardSplit(
-            dataset_name=DatasetName.PASSION,
-            config=_config,
-            SSL_model=model,
-            append_to_df=args.append_results,
-            log_wandb=log_wandb,
-            add_info="impetigo",
-        )
-        trainer.evaluate()
