+++ src/models/helper.py
@@ -10,15 +10,11 @@ from tqdm.auto import tqdm
 ARR_TYPE = Union[np.ndarray, np.memmap, torch.Tensor]
 
 
-# todo: add performance improvements to main branch
 def embed_dataset(
     torch_dataset: torch.utils.data.DataLoader,
     model: Optional[torch.nn.Sequential],
     n_layers: int,
     normalize: bool = False,
-    # memmap: bool = True,
-    # memmap_path: Union[Path, str, None] = None,
-    return_only_embedding_and_labels: bool = False,
     tqdm_desc: Optional[str] = None,
 ) -> Union[Tuple[ARR_TYPE, ARR_TYPE, ARR_TYPE, ARR_TYPE], Tuple[ARR_TYPE, ARR_TYPE]]:
     labels = []
@@ -46,31 +42,12 @@ def embed_dataset(
         emb_dim = _batch.squeeze().shape[0]
     else:
         emb_dim = model(_batch, n_layers=n_layers).squeeze().shape[0]
-    # create the memmap's
-    # if memmap:
-    #     memmap_path = create_memmap_path(memmap_path=memmap_path)
-    #     emb_space = create_memmap(
-    #         memmap_path,
-    #         "embedding_space.dat",
-    #         len(torch_dataset.dataset),
-    #         *(emb_dim,),
-    #     )
-    #     # if not return_only_embedding_and_labels:
-    #     images = create_memmap(
-    #         memmap_path,
-    #         "images.dat",
-    #         len(torch_dataset.dataset),
-    #         *batch_dim,
-    #     )
-    # else:
+
     emb_space = np.zeros(shape=(len(torch_dataset.dataset), emb_dim))
-    # if not return_only_embedding_and_labels:
     images = np.zeros(shape=(len(torch_dataset.dataset), *batch_dim))
-    # else done
 
     del emb_dim, batch_dim, _batch
     # embed the dataset
-
     if (
         type(model) is torch.jit._script.RecursiveScriptModule
         or type(model) is torch.nn.Sequential
@@ -82,11 +59,6 @@ def embed_dataset(
     for i, batch_tup in iterator:
         if len(batch_tup) == 4:
             batch, path, label, index = batch_tup
-        # elif len(batch_tup) == 3:
-        #    batch, path, label = batch_tup
-        # elif len(batch_tup) == 2:
-        #    batch, label = batch_tup
-        #    path = None
         else:
             raise ValueError(f"Unknown batch tuple: {batch_tup}")
 
@@ -104,52 +76,15 @@ def embed_dataset(
             if normalize:
                 emb = torch.nn.functional.normalize(emb, dim=-1, p=2)
             emb_space[batch_size * i : batch_size * (i + 1), :] = emb.cpu()
-            # if memmap:
-            #    emb_space.flush()
             labels.append(label.cpu())
-            # if not return_only_embedding_and_labels:
             images[batch_size * i : batch_size * (i + 1), :] = batch.cpu()
-            # if memmap:
-            #    images.flush()
             if path is not None:
                 paths += path
             if index is not None:
                 indices += index
     labels = torch.concat(labels).cpu()
 
-    # if return_only_embedding_and_labels:
-    #    return emb_space, labels
-    # if len(paths) > 0:
     paths = np.array(paths)
-    # else:
-    #    paths = None
-    # if len(indices) > 0:
     indices = np.array(indices)
-    # else:
-    #    indices = None
 
     return emb_space, labels, images, paths, indices
-
-
-def create_memmap(memmap_path: Path, memmap_file_name: str, len_dataset: int, *dims):
-    memmap_file = memmap_path / memmap_file_name
-    if memmap_file.exists():
-        memmap_file.unlink()
-    memmap = np.memmap(
-        str(memmap_file),
-        dtype=np.float32,
-        mode="w+",
-        shape=(len_dataset, *dims),
-    )
-    return memmap
-
-
-def create_memmap_path(memmap_path: Union[str, Path, None]) -> Path:
-    if memmap_path is None:
-        # temporary folder for saving memory map
-        memmap_path = Path(tempfile.mkdtemp())
-    else:
-        # make sure the path exists
-        memmap_path = Path(memmap_path)
-        memmap_path.mkdir(parents=True, exist_ok=True)
-    return memmap_path
